Removed view cone and minor refactor
