+++ test/algorithm/test_visibility_finder.py
@@ -5,10 +5,10 @@ from kaos.algorithm.visibility_finder import VisibilityFinder
 from kaos.models import Satellite
 from kaos.models.parser import parse_ephemeris_file
 
-from . import KaosVisibilityFinderTestCase
+from .. import KaosTestCase
 
 @ddt
-class TestVisibilityFinder(KaosVisibilityFinderTestCase):
+class TestVisibilityFinder(KaosTestCase):
     """Test the visibility finder's accuracy."""
 
     @classmethod
@@ -16,11 +16,11 @@ class TestVisibilityFinder(KaosVisibilityFinderTestCase):
         super(TestVisibilityFinder, cls).setUpClass()
         parse_ephemeris_file("ephemeris/Radarsat2.e")
 
-    @data(('test/algorithm/vancouver.test', (1514764802, 1514772000), 60),
-          ('test/algorithm/vancouver.test', (1514768543, 1514772000), 60),
-          ('test/algorithm/vancouver.test', (1514768340, 1514768400), 60),
-          ('test/algorithm/vancouver.test', (1514768543, 1514769143), 60),
-          ('test/algorithm/vancouver.test', (1515160800, 1515164400), 60))
+    @data(('test/test_data/vancouver.test', (1514764802, 1514772000), 60),
+          ('test/test_data/vancouver.test', (1514768543, 1514772000), 60),
+          ('test/test_data/vancouver.test', (1514768340, 1514768400), 60),
+          ('test/test_data/vancouver.test', (1514768543, 1514769143), 60),
+          ('test/test_data/vancouver.test', (1515160800, 1515164400), 60))
     def test_full_visibility(self, test_data):
         """Tests that the visibility finder produces the same results as the access file.
 
@@ -66,4 +66,3 @@ class TestVisibilityFinder(KaosVisibilityFinderTestCase):
 
             if not found:
                 raise Exception('Wrong access: {}'.format(predicted_access))
-
