--- test/algorithm/vancouver.test
@@ -1,95 +1,0 @@
-                                KMC ACCESS TIME TEST FILE
-====================================================================================================
-Satellite Name: Radarsat2
-Target Point: 49.07, -123.113
-====================================================================================================
-   1,    1 Jan 2018 00:58:52.849,    1 Jan 2018 01:13:45.407,          892.558
-   2,    1 Jan 2018 02:39:16.290,    1 Jan 2018 02:53:06.257,          829.967
-   3,    1 Jan 2018 13:07:57.129,    1 Jan 2018 13:20:30.468,          753.340
-   4,    1 Jan 2018 14:46:54.910,    1 Jan 2018 15:02:00.687,          905.777
-   5,    1 Jan 2018 16:27:00.381,    1 Jan 2018 16:39:33.366,          752.985
-   6,    1 Jan 2018 18:08:48.153,    1 Jan 2018 18:12:32.917,          224.764
-   7,    1 Jan 2018 22:54:52.292,    1 Jan 2018 23:03:56.974,          544.682
-   8,    2 Jan 2018 00:30:19.251,    2 Jan 2018 00:44:35.323,          856.072
-   9,    2 Jan 2018 02:09:33.525,    2 Jan 2018 02:24:16.000,          882.475
-  10,    2 Jan 2018 03:53:49.255,    2 Jan 2018 04:01:52.910,          483.655
-  11,    2 Jan 2018 12:39:37.078,    2 Jan 2018 12:49:47.039,          609.961
-  12,    2 Jan 2018 14:17:53.327,    2 Jan 2018 14:32:51.771,          898.444
-  13,    2 Jan 2018 15:57:43.424,    2 Jan 2018 16:11:28.955,          825.531
-  14,    2 Jan 2018 17:38:36.958,    2 Jan 2018 17:46:12.420,          455.462
-  15,    2 Jan 2018 22:27:58.900,    2 Jan 2018 22:34:10.836,          371.937
-  16,    3 Jan 2018 00:02:04.087,    3 Jan 2018 00:15:20.892,          796.805
-  17,    3 Jan 2018 01:40:13.129,    3 Jan 2018 01:55:17.787,          904.659
-  18,    3 Jan 2018 03:22:33.423,    3 Jan 2018 03:33:54.988,          681.565
-  19,    3 Jan 2018 12:12:12.558,    3 Jan 2018 12:17:49.319,          336.761
-  20,    3 Jan 2018 13:48:58.132,    3 Jan 2018 14:03:21.982,          863.849
-  21,    3 Jan 2018 15:28:30.816,    3 Jan 2018 15:43:05.641,          874.825
-  22,    3 Jan 2018 17:09:00.001,    3 Jan 2018 17:19:03.580,          603.579
-  23,    3 Jan 2018 23:34:08.209,    3 Jan 2018 23:46:01.744,          713.536
-  24,    4 Jan 2018 01:11:12.665,    4 Jan 2018 01:26:13.913,          901.248
-  25,    4 Jan 2018 02:52:08.110,    4 Jan 2018 03:05:24.280,          796.170
-  26,    4 Jan 2018 13:20:12.132,    4 Jan 2018 13:33:27.841,          795.710
-  27,    4 Jan 2018 14:59:22.328,    4 Jan 2018 15:14:23.791,          901.463
-  28,    4 Jan 2018 16:39:34.389,    4 Jan 2018 16:51:28.682,          714.294
-  29,    4 Jan 2018 23:06:33.434,    4 Jan 2018 23:16:36.071,          602.637
-  30,    5 Jan 2018 00:42:30.901,    5 Jan 2018 00:57:05.387,          874.486
-  31,    5 Jan 2018 02:22:14.103,    5 Jan 2018 02:36:38.198,          864.096
-  32,    5 Jan 2018 04:07:45.311,    5 Jan 2018 04:13:24.808,          339.496
-  33,    5 Jan 2018 12:51:41.571,    5 Jan 2018 13:03:02.264,          680.692
-  34,    5 Jan 2018 14:30:18.500,    5 Jan 2018 14:45:23.187,          904.688
-  35,    5 Jan 2018 16:10:15.299,    5 Jan 2018 16:23:32.657,          797.357
-  36,    5 Jan 2018 17:51:25.037,    5 Jan 2018 17:57:38.490,          373.453
-  37,    5 Jan 2018 22:39:24.812,    5 Jan 2018 22:46:59.018,          454.207
-  38,    6 Jan 2018 00:14:07.713,    6 Jan 2018 00:27:52.781,          825.068
-  39,    6 Jan 2018 01:52:44.464,    6 Jan 2018 02:07:42.980,          898.516
-  40,    6 Jan 2018 03:35:48.462,    6 Jan 2018 03:45:59.601,          611.138
-  41,    6 Jan 2018 12:23:43.916,    6 Jan 2018 12:31:45.858,          481.942
-  42,    6 Jan 2018 14:01:20.248,    6 Jan 2018 14:16:02.553,          882.305
-  43,    6 Jan 2018 15:41:00.803,    6 Jan 2018 15:55:17.268,          856.465
-  44,    6 Jan 2018 17:21:38.991,    6 Jan 2018 17:30:44.715,          545.724
-  45,    6 Jan 2018 22:13:04.811,    6 Jan 2018 22:16:47.180,          222.369
-  46,    6 Jan 2018 23:46:03.297,    6 Jan 2018 23:58:35.682,          752.386
-  47,    7 Jan 2018 01:23:35.533,    7 Jan 2018 01:38:41.249,          905.716
-  48,    7 Jan 2018 03:05:05.178,    7 Jan 2018 03:17:39.237,          754.059
-  49,    7 Jan 2018 13:32:29.922,    7 Jan 2018 13:46:19.498,          829.576
-  50,    7 Jan 2018 15:11:50.628,    7 Jan 2018 15:26:43.432,          892.804
-  51,    7 Jan 2018 16:52:09.585,    7 Jan 2018 17:03:20.129,          670.544
-  52,    7 Jan 2018 23:18:19.241,    7 Jan 2018 23:29:13.282,          654.041
-  53,    8 Jan 2018 00:54:45.957,    8 Jan 2018 01:09:34.764,          888.807
-  54,    8 Jan 2018 02:34:59.121,    8 Jan 2018 02:48:58.805,          839.684
-  55,    8 Jan 2018 13:03:51.628,    8 Jan 2018 13:16:08.665,          737.037
-  56,    8 Jan 2018 14:42:44.716,    8 Jan 2018 14:57:50.945,          906.229
-  57,    8 Jan 2018 16:22:48.016,    8 Jan 2018 16:35:32.833,          764.818
-  58,    8 Jan 2018 18:04:22.983,    8 Jan 2018 18:08:52.217,          269.234
-  59,    8 Jan 2018 22:50:58.448,    8 Jan 2018 22:59:42.201,          523.753
-  60,    9 Jan 2018 00:26:14.850,    9 Jan 2018 00:40:23.932,          849.082
-  61,    9 Jan 2018 02:05:19.654,    9 Jan 2018 02:20:07.070,          887.416
-  62,    9 Jan 2018 03:49:15.383,    9 Jan 2018 03:57:55.818,          520.436
-  63,    9 Jan 2018 12:35:36.009,    9 Jan 2018 12:45:18.274,          582.265
-  64,    9 Jan 2018 14:13:43.884,    9 Jan 2018 14:28:39.244,          895.360
-  65,    9 Jan 2018 15:53:31.668,    9 Jan 2018 16:07:25.726,          834.059
-  66,    9 Jan 2018 17:34:20.847,    9 Jan 2018 17:42:20.693,          479.846
-  67,    9 Jan 2018 22:24:11.317,    9 Jan 2018 22:29:51.713,          340.397
-  68,    9 Jan 2018 23:58:02.386,   10 Jan 2018 00:11:08.793,          786.407
-  69,   10 Jan 2018 01:36:02.237,   10 Jan 2018 01:51:07.863,          905.626
-  70,   10 Jan 2018 03:18:09.208,   10 Jan 2018 03:29:50.827,          701.618
-  71,   10 Jan 2018 12:08:31.177,   10 Jan 2018 12:12:57.919,          266.742
-  72,   10 Jan 2018 13:44:49.758,   10 Jan 2018 13:59:06.227,          856.469
-  73,   10 Jan 2018 15:24:19.623,   10 Jan 2018 15:38:59.746,          880.123
-  74,   10 Jan 2018 17:04:46.078,   10 Jan 2018 17:15:07.545,          621.468
-  75,   10 Jan 2018 23:30:09.303,   10 Jan 2018 23:41:48.877,          699.575
-  76,   11 Jan 2018 01:07:04.425,   11 Jan 2018 01:22:03.277,          898.852
-  77,   11 Jan 2018 02:47:48.985,   11 Jan 2018 03:01:17.302,          808.317
-  78,   11 Jan 2018 13:16:05.532,   11 Jan 2018 13:29:08.085,          782.554
-  79,   11 Jan 2018 14:55:11.801,   11 Jan 2018 15:10:15.192,          903.391
-  80,   11 Jan 2018 16:35:21.616,   11 Jan 2018 16:47:29.404,          727.788
-  81,   11 Jan 2018 18:17:58.998,   11 Jan 2018 18:19:25.409,           86.411
-  82,   11 Jan 2018 23:02:37.859,   11 Jan 2018 23:12:22.030,          584.171
-  83,   12 Jan 2018 00:38:25.338,   12 Jan 2018 00:52:54.159,          868.821
-  84,   12 Jan 2018 02:17:58.800,   12 Jan 2018 02:32:29.664,          870.864
-  85,   12 Jan 2018 04:03:01.295,   12 Jan 2018 04:09:36.448,          395.153
-  86,   12 Jan 2018 12:47:38.027,   12 Jan 2018 12:58:36.902,          658.875
-  87,   12 Jan 2018 14:26:08.625,   12 Jan 2018 14:41:11.778,          903.153
-  88,   12 Jan 2018 16:06:03.191,   12 Jan 2018 16:19:30.484,          807.293
-  89,   12 Jan 2018 17:47:06.799,   12 Jan 2018 17:53:49.574,          402.775
-  90,   12 Jan 2018 22:35:33.786,   12 Jan 2018 22:42:42.335,          428.549
