+++ test/api/test_visibility.py
@@ -3,8 +3,9 @@
 
 from collections import namedtuple
 import re
+import logging
 
-from ddt import ddt, data
+from ddt import ddt, data, file_data
 
 from kaos.algorithm.visibility_finder import VisibilityFinder
 from kaos.models import Satellite
@@ -13,7 +14,6 @@ from kaos.utils.time_conversion import utc_to_unix
 
 from .. import KaosTestCase
 
-AccessTestInfo = namedtuple('AccessTestInfo', 'sat_name, target, accesses')
 
 @ddt
 class TestVisibilityApi(KaosTestCase):
@@ -25,57 +25,10 @@ class TestVisibilityApi(KaosTestCase):
         super(TestVisibilityApi, cls).setUpClass()
         parse_ephemeris_file("ephemeris/Radarsat2.e")
 
-    # pylint: disable=line-too-long
-    @staticmethod
-    def parse_access_file(file_path, access_range=None):
-        """Reads a KAOS access test file, these files follow the following format:
-
-            ====================================================================================================
-            Satellite Name: <Sat Name>
-            Target Point: <lon>, <lat>
-            ====================================================================================================
-            record number, access start, access_end, access_duration
-            ....
-
-        Args:
-            file_path (string): The path of the KAOS access test file.
-            access_range (tuple):
-
-        Returns:
-            An AccessTestInfo tuple.
-        """
-        with open(file_path) as access_file:
-            access_info_text = access_file.read()
-
-        section_regex = re.compile(r'={99}', re.MULTILINE)
-        access_info = section_regex.split(access_info_text)
-
-        # Parse the header
-        sat_name = re.search(r'Satellite Name: ([a-zA-Z0-9]+)', access_info[1]).groups()[0]
-        target = [float(point) for point in
-                  re.search(r'Target Point: (.*)', access_info[1]).groups()[0].split(',')]
-        # Parse the access times
-        accesses = []
-        raw_access_data = access_info[2].splitlines()
-        for access in raw_access_data[1:]:
-            access = access.split(',')
-            # Parse the start and end time
-            start_time = utc_to_unix(access[1].rstrip().lstrip(), '%d %b %Y %H:%M:%S.%f')
-            end_time = utc_to_unix(access[2].rstrip().lstrip(), '%d %b %Y %H:%M:%S.%f')
-            if (access_range is None
-                or ((start_time >= access_range[0] and start_time <= access_range[1]) or
-                    (end_time >= access_range[0] and end_time <= access_range[1]) or
-                    (start_time <= access_range[0] and end_time >= access_range[0]) or
-                    (end_time <= access_range[1] and end_time >= access_range[1]))):
-                accesses.append((start_time, end_time))
-
-        return AccessTestInfo(sat_name, target, accesses)
-    # pylint: enable=line-too-long
-
-    @data(('test/algorithm/vancouver.test', ('20180101T00:00:00.0', '20180101T02:00:00.0'), 60),
-          ('test/algorithm/vancouver.test', ('20180104T00:00:00.0', '20180107T11:59:59.0'), 60),
-          ('test/algorithm/vancouver.test', ('20180102T00:00:00.0', '20180102T11:59:59.0'), 60))
-    def test_full_visibility(self, test_data):
+    @data(('test/test_data/vancouver.test', ('20180101T00:00:00.0', '20180101T02:00:00.0'), 60),
+          ('test/test_data/vancouver.test', ('20180104T00:00:00.0', '20180107T11:59:59.0'), 60),
+          ('test/test_data/vancouver.test', ('20180102T00:00:00.0', '20180102T11:59:59.0'), 60))
+    def test_visibility(self, test_data):
         """Tests that the visibility finder produces the same results as the access file.
 
         Args:
@@ -134,3 +87,7 @@ class TestVisibilityApi(KaosTestCase):
 
             if not found:
                 raise Exception('Wrong access: {}'.format(predicted_access))
+
+    @file_data("test_data_visibility.json")
+    def test_visibility_incorrect_input(self, data):
+        logging.info(data)
